requirements.txt
streamlit
requests
beautifulsoup4
langchain
langchain-community
sentence-transformers
chromadb
python-dotenv
tqdm